feat(auth): implemented verify email page. Closes #673
