+++ _TO_CONVERT_NOW/0example-react-recent/Layouts/AppLayout.vue
@@ -8,10 +8,6 @@ import DropdownLink from '@/Components/DropdownLink.vue';
 import NavLink from '@/Components/NavLink.vue';
 import ResponsiveNavLink from '@/Components/ResponsiveNavLink.vue';
 
-defineProps({
-  title: String,
-});
-
 const showingNavigationDropdown = ref(false);
 
 const switchToTeam = (team) => {
