+++ apps/knowii/src/Pages/Auth/ConfirmPassword.tsx
@@ -2,10 +2,10 @@ import FooterGuest from '@/Components/FooterGuest';
 import { Head, useForm } from '@inertiajs/react';
 import React, { FormEventHandler } from 'react';
 import { useRoute } from 'ziggy-js';
-import { EMAIL_PASSWORD_URL, PASSWORD_CONFIRM_URL } from '@knowii/common';
+import { PASSWORD_CONFIRM_URL } from '@knowii/common';
 import AuthenticationCard from '@/Components/AuthenticationCard';
 import InputLabel from '@/Components/InputLabel';
-import { FaAt, FaLock } from 'react-icons/fa';
+import { FaLock } from 'react-icons/fa';
 import { InputText } from 'primereact/inputtext';
 import InputError from '@/Components/InputError';
 import { Button } from 'primereact/button';
